opensuse updates and fixes

+++ config/niri/pipewire.sh
@@ -1,11 +1,16 @@
 #!/usr/bin/env bash
 
-ITEMS=("pipewire" "pipewire-pulse" "wireplumber")
+if command -v systemctl;
+then
+    # systemd should handle this
+else
+    ITEMS=("pipewire" "pipewire-pulse" "wireplumber")
 
-for item in ${ITEMS[@]}; do
-    killall -9 $item
-done
+    for item in ${ITEMS[@]}; do
+        killall -9 $item
+    done
 
-for item in ${ITEMS[@]}; do
-    $item &
-done
+    for item in ${ITEMS[@]}; do
+        $item &
+    done
+fi
